+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Fix the issue that the option "Highlight Troops When Showing Indicators" is always reset to default on game start.
 
+- Fix potential crash with other mods that creates new custom units.
+
 ### Improved
 - After dragging mouse right button to rotate camera, mouse position will be recovered to the position before dragging, instead of to the center of window.